+++ src/tools/mod.rs
@@ -1,6 +1,7 @@
 pub mod affinity;
 pub mod animation;
 pub mod asset_audit;
+pub mod attributes;
 pub mod character;
 pub mod core;
 pub mod datastore;
@@ -20,6 +21,7 @@ pub mod profiler;
 pub mod profiler_v2;
 pub mod publish;
 pub mod scenario;
+pub mod screenshot;
 pub mod script_patch;
 pub mod scripts;
 pub mod security;